migrate 20x4 lcd controller to class
